requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
workalendar==17.0.0